Backend/requirements.txt
# ============================================================
# Finance Buddy Backend - Pinned Requirements
# ALL versions are explicitly pinned to PREVENT dependency
# backtracking. Update versions carefully and systematically.
# ============================================================

# --- Core Framework ---
fastapi==0.135.1
uvicorn==0.41.0
starlette==0.52.1

# --- Data Validation ---
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
annotated-types==0.7.0

# --- Database ---
motor==3.7.1
pymongo==4.16.0
dnspython==2.8.0

# --- Auth & Security ---
bcrypt==4.3.0
python-jose==3.5.0
cryptography==46.0.5
cffi==2.0.0
pycparser==3.0
ecdsa==0.19.1
rsa==4.9.1
pyasn1==0.6.2
pyasn1_modules==0.4.2

# --- Utilities ---
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.2
email-validator==2.3.0
packaging==25.0
typing_extensions==4.15.0
typing-inspection==0.4.2
typing-inspect==0.9.0

# --- LangChain Core Stack ---
langchain==0.3.25
langchain-core==0.3.83
langchain-community==0.3.23
langchain-text-splitters==0.3.11
langsmith==0.3.45

# --- LangChain Integrations ---
langchain-chroma==0.2.2
langchain-google-genai==2.0.9
langchain-ollama==0.3.2

# --- ChromaDB & Vector Store ---
chromadb==0.6.3
chroma-hnswlib==0.7.6
onnxruntime==1.24.4
tokenizers==0.22.2
mmh3==5.2.1
PyPika==0.51.1
overrides==7.7.0
importlib_resources==6.5.2
importlib_metadata==8.7.1
zipp==3.23.0
zstandard==0.23.0

# --- Google AI ---
filetype==1.2.0
google-generativeai==0.8.6
google-genai>=1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
googleapis-common-protos==1.73.0
proto-plus==1.27.1
protobuf==5.29.6
grpcio==1.78.0
grpcio-status==1.71.2
httplib2==0.31.2
uritemplate==4.2.0

# --- Ollama ---
ollama==0.6.1

# --- Sentence Transformers & ML ---
sentence-transformers==3.3.1
torch==2.10.0
transformers==4.57.6
huggingface_hub==0.36.2
safetensors==0.7.0
tokenizers==0.22.2
scikit-learn==1.8.0
scipy==1.17.1
numpy==1.26.4
joblib==1.5.3
threadpoolctl==3.6.0
pillow==12.1.1
sympy==1.14.0
mpmath==1.3.0
networkx==3.6.1
filelock==3.25.2
fsspec==2026.2.0
regex==2026.2.28
tqdm==4.67.3
setuptools==82.0.1

# --- HTTP & Networking ---
httpx==0.28.1
httpx-sse==0.4.3
httptools==0.7.1
httpcore==1.0.9
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
aiohttp==3.13.3
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
frozenlist==1.8.0
multidict==6.7.1
yarl==1.23.0
propcache==0.4.1
oauthlib==3.3.1
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
urllib3==2.6.3
anyio==4.12.1
h11==0.16.0

# --- OpenTelemetry (ChromaDB observability) ---
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-asgi==0.61b0
opentelemetry-instrumentation-fastapi==0.61b0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-util-http==0.61b0

# --- Misc Dependencies ---
pypdf==5.1.0
SQLAlchemy==2.0.48
greenlet==3.3.2
dataclasses-json==0.6.7
marshmallow==3.26.2
mypy_extensions==1.1.0
tenacity==9.1.4
jsonpatch==1.33
jsonpointer==3.0.0
uuid_utils==0.14.1
orjson==3.11.7
backoff==2.2.1
posthog==7.9.12
distro==1.9.0
Jinja2==3.1.6
MarkupSafe==3.0.3
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2
rich==14.3.3
typer==0.24.1
shellingham==1.5.4
click==8.3.1
colorama==0.4.6
six==1.17.0
asgiref==3.11.1
websocket-client==1.9.0
websockets==16.0
watchfiles==1.1.1
python-dateutil==2.9.0.post0
pyparsing==3.3.2
pyproject_hooks==1.2.0
build==1.4.0
wrapt==1.17.3
kubernetes==35.0.0
durationpy==0.10
flatbuffers==25.12.19
attrs==25.4.0
annotated-doc==0.0.4
